remove link class from view model service
