+++ nav-app/src/app/classes/index.ts
@@ -15,3 +15,4 @@ export { ContextMenuItem } from "./context-menu-item";
 export { Gradient, Gcolor} from "./gradient";
 export { Filter } from "./filter";
 export { LayoutOptions } from "./layout-options";
+export { Link } from "./link";
