+++ nav-app/src/app/services/viewmodels.service.ts
@@ -1,5 +1,5 @@
 import { EventEmitter, Injectable, Output } from '@angular/core';
-import { Technique, Tactic, Matrix, VersionChangelog, Gradient, Filter, LayoutOptions } from '../classes';
+import { Technique, Tactic, Matrix, VersionChangelog, Gradient, Filter, LayoutOptions, Link } from '../classes';
 import { DataService } from './data.service';
 import * as tinycolor from 'tinycolor2';
 import { evaluate } from 'mathjs';
@@ -1623,41 +1623,5 @@ export class Metadata {
 
     valid(): boolean {
         return (this.name && this.name.length > 0 && this.value && this.value.length > 0) || (this.divider !== undefined)
-    }
-}
-
-// { label, url } with serialization
-export class Link {
-    public label: string;
-    public url: string;
-    public divider: boolean;
-
-    constructor() { }
-
-    serialize(): object { 
-        return this.label && this.url ? {label: this.label, url: this.url} : {divider: this.divider};
-    }
-
-    deSerialize(rep: any): void {
-        let obj = (typeof(rep) == "string")? JSON.parse(rep) : rep;
-        if ("url" in obj) { // label & url object
-            if (typeof(obj.url) === "string") this.url = obj.url;
-            else console.error("TypeError: Link field 'url' is not a string");
-
-            if ("label" in obj) {
-                if (typeof(obj.label) === "string") this.label = obj.label;
-                else console.error("TypeError: Link field 'label' is not a string");
-            }
-            else console.error("Error: Link required field 'label' not present");
-        }
-        else if ("divider" in obj) { // divider object
-            if (typeof(obj.divider) === "boolean") this.divider = obj.divider;
-            else  console.error("TypeError: Link field 'divider' is not a boolean");
-        }
-        else console.error("Error: Link required field 'url' or 'divider' not present");
-    }
-
-    valid(): boolean {
-        return (this.label && this.label.length > 0 && this.url && this.url.length > 0) || (this.divider !== undefined)
     }
 }
